Refactor Signed-off-by: Alexander Tkachev <atkachev@gmail.com>
